🧑‍💻 Split up header by class
