+++ src/source/DRVCTRL.cpp
@@ -3,7 +3,7 @@
  * DRVCTRL.cpp - Driver Control
  * TMC2660
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 #define SET_REG0(SETTING) DRVCTRL_0_register.SETTING = B; DRVCTRL(DRVCTRL_0_register.sr)
