+++ src/source/DRVSTATUS.cpp
@@ -3,7 +3,7 @@
  * DRVSTATUS.cpp - Driver Status
  * TMC2660
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 #define GET_REG00(SETTING) DRVSTATUS(); return READ_RDSEL00_register.SETTING
