+++ src/source/PWMCONF.cpp
@@ -3,7 +3,7 @@
  * PWMCONF.cpp - PWM Configuration
  * TMC2130 (TMC2160, TMC5130, TMC5160, TMC5161), TMC2160 (TMC5130, TMC5160, TMC5161), TMC2208 (TMC2209, TMC2224)
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 #define SET_REG(SETTING) PWMCONF_register.SETTING = B; write(PWMCONF_register.address, PWMCONF_register.sr)
