+++ src/source/RAMP_STAT.cpp
@@ -3,7 +3,7 @@
  * RAMP_STAT.cpp - Ramp Status
  * TMC5130 (TMC5160, TMC5161)
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 #define GET_REG(SETTING) RAMP_STAT_t r{0}; r.sr = RAMP_STAT(); return r.SETTING
