+++ src/source/SGCSCONF.cpp
@@ -3,7 +3,7 @@
  * SGCSCONF.cpp - SGCS Configuration
  * TMC2660
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 #define SET_REG(SETTING) SGCSCONF_register.SETTING = B; write(SGCSCONF_register.address, SGCSCONF_register.sr);
