+++ src/source/SLAVECONF.cpp
@@ -3,7 +3,7 @@
  * SLAVECONF.cpp
  * TMC2208 (TMC2209, TMC2224), TMC2240, TMC5130
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 #define SET_REG(SETTING) SLAVECONF_register.SETTING = B; write(SLAVECONF_register.address, SLAVECONF_register.sr)
@@ -30,6 +30,8 @@ void TMC2240Stepper::SLAVECONF(uint16_t input) {
 	SLAVECONF_register.sr = input & 0xF00;
 	write(SLAVECONF_register.address, SLAVECONF_register.sr);
 }
+uint8_t TMC2240Stepper::slaveaddr()			{ GET_REG(slaveaddr); }
+void TMC2240Stepper::slaveaddr(uint8_t B)	{ SET_REG(slaveaddr); }
 uint8_t TMC2240Stepper::senddelay()			{ GET_REG(senddelay); }
 void TMC2240Stepper::senddelay(uint8_t B)	{ SET_REG(senddelay); }
