+++ src/source/SMARTEN.cpp
@@ -3,7 +3,7 @@
  * SMARTEN.cpp - Smart Enable
  * TMC2660
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 #define SET_REG(SETTING) SMARTEN_register.SETTING = B; write(SMARTEN_register.address, SMARTEN_register.sr)
