+++ src/source/TMC2160Stepper.cpp
@@ -3,7 +3,7 @@
  * TMC2160Stepper.cpp
  * Implementing methods for TMC2160 (TMC5130, TMC5160, TMC5161)
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 TMC2160Stepper::TMC2160Stepper(uint16_t pinCS, float RS, int8_t link) : TMC2130Stepper(pinCS, RS, link)
