+++ src/source/TMC5130Stepper.cpp
@@ -3,7 +3,7 @@
  * TMC5130Stepper.cpp
  * Implementing methods for TMC5130 (TMC5160, TMC5161)
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 TMC5130Stepper::TMC5130Stepper(uint16_t pinCS, float RS, int8_t link) : TMC2160Stepper(pinCS, RS, link)
