+++ src/source/TMC5160Stepper.cpp
@@ -3,7 +3,7 @@
  * TMC5160Stepper.cpp
  * Implementing methods for TMC5160 (TMC5161)
  */
-#include "TMCStepper.h"
+#include "../TMCStepper.h"
 #include "TMC_MACROS.h"
 
 TMC5160Stepper::TMC5160Stepper(uint16_t pinCS, float RS, int8_t link) : TMC5130Stepper(pinCS, RS, link)
