+++ src/source/TMCStepperBase.h
@@ -1,0 +1,92 @@
+/**
+ * TMCStepper library by @teemuatlut
+ *
+ * TMCStepperBase.h
+ * Individual header for class TMCStepper
+ */
+#pragma once
+
+class TMCStepper {
+	public:
+		uint16_t cs2rms(uint8_t CS);
+		void rms_current(uint16_t mA);
+		void rms_current(uint16_t mA, float mult);
+		uint16_t rms_current();
+		void hold_multiplier(float val) { holdMultiplier = val; }
+		float hold_multiplier() { return holdMultiplier; }
+		uint8_t test_connection();
+
+		// Helper functions
+		void microsteps(uint16_t ms);
+		uint16_t microsteps();
+		void blank_time(uint8_t value);
+		uint8_t blank_time();
+		void hysteresis_end(int8_t value);
+		int8_t hysteresis_end();
+		void hysteresis_start(uint8_t value);
+		uint8_t hysteresis_start();
+
+		// R+WC: GSTAT
+		void 	GSTAT(							uint8_t input);
+		uint8_t GSTAT();
+		bool 	reset();
+		bool 	drv_err();
+		bool 	uv_cp();
+
+		// W: IHOLD_IRUN
+		void IHOLD_IRUN(					uint32_t input);
+		uint32_t IHOLD_IRUN();
+		void 	ihold(							uint8_t B);
+		void 	irun(								uint8_t B);
+		void 	iholddelay(					uint8_t B);
+		uint8_t ihold();
+		uint8_t irun();
+		uint8_t iholddelay();
+
+		// W: TPOWERDOWN
+		uint8_t TPOWERDOWN();
+		void TPOWERDOWN(					uint8_t input);
+
+		// R: TSTEP
+		uint32_t TSTEP();
+
+		// W: TPWMTHRS
+		uint32_t TPWMTHRS();
+		void TPWMTHRS(						uint32_t input);
+
+		// R: MSCNT
+		uint16_t MSCNT();
+
+		// R: MSCURACT
+		uint32_t MSCURACT();
+		int16_t cur_a();
+		int16_t cur_b();
+
+	protected:
+		TMCStepper(float RS) : Rsense(RS) {};
+		INIT_REGISTER(IHOLD_IRUN){{.sr=0}};	// 32b
+		INIT_REGISTER(TPOWERDOWN){.sr=0};		// 8b
+		INIT_REGISTER(TPWMTHRS){.sr=0};			// 32b
+
+		static constexpr uint8_t TMC_READ = 0x00, TMC_WRITE = 0x80;
+
+		struct TSTEP_t { constexpr static uint8_t address = 0x12; };
+		struct MSCNT_t { constexpr static uint8_t address = 0x6A; };
+
+		virtual void write(uint8_t, uint32_t) = 0;
+		virtual uint32_t read(uint8_t) = 0;
+		virtual void vsense(bool) = 0;
+		virtual bool vsense(void) = 0;
+		virtual uint32_t DRV_STATUS() = 0;
+		virtual void hend(uint8_t) = 0;
+		virtual uint8_t hend() = 0;
+		virtual void hstrt(uint8_t) = 0;
+		virtual uint8_t hstrt() = 0;
+		virtual void mres(uint8_t) = 0;
+		virtual uint8_t mres() = 0;
+		virtual void tbl(uint8_t) = 0;
+		virtual uint8_t tbl() = 0;
+
+		const float Rsense;
+		float holdMultiplier = 0.5;
+};
